feat: logo uploader for individual stores
This feature is about adding the functionality of adding a logo for every store. It adds a small compression on the logo so make it smaller and converts it to webp for better performance. Additionally it sends it to UploadThing and after that it is added to the store collection